refactor(Approvals): organising standardised directories for API consuming services
